+++ packages/flame/CHANGELOG.md
@@ -56,6 +56,7 @@
  - `HasGameRef` can now operate independently from `Game`
  - `initialDelay` and `peakDelay` for effects to handle time before and after an effect
  - Removes `component.onMount` since `onLoad` can be used for all its purposes
+ - Add `component.onParentChange` which runs after `component.prepare` and `component.onLoad`
 
 ## [1.0.0-releasecandidate.13]
  - Fix camera not ending up in the correct position on long jumps